--- public/style.css
@@ -1,22 +1,0 @@
-html {
-    scroll-padding-top: 70px;
-}
-.container {
-    margin-top: 60px;
-}
-#searchresult {
-    display: none;
-    width: 400px;
-    min-height: 125px;
-    background-color: #333;
-    position: fixed;
-    top: 60px;
-    border-radius: 6px;
-    padding: 5px;
-    /* right: 20px; */
-    color: white;
-}
-
-#searchresult a {
-    color: white;
-}
